requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
coverage==5.5
Flask==2.0.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
PyYAML==5.4.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.4
Werkzeug==2.0.0
